+++ src/main/java/org/quiltmc/qsl/frozenblock/core/registry/impl/sync/ServerPackets.java
@@ -1,0 +1,155 @@
+/*
+ * Copyright 2022 The Quilt Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.quiltmc.qsl.frozenblock.core.registry.impl.sync;
+
+import java.util.Collection;
+
+import it.unimi.dsi.fastutil.ints.IntList;
+import net.fabricmc.fabric.api.networking.v1.FabricPacket;
+import net.fabricmc.fabric.api.networking.v1.PacketType;
+import net.frozenblock.lib.FrozenMain;
+import net.minecraft.network.FriendlyByteBuf;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
+import net.minecraft.resources.ResourceLocation;
+import org.jetbrains.annotations.ApiStatus;
+import org.quiltmc.qsl.frozenblock.core.registry.impl.sync.mod_protocol.ModProtocolDef;
+
+/**
+ * Identifiers of packets sent by server.
+ */
+@ApiStatus.Internal
+public final class ServerPackets {
+	/**
+	 * Starts registry sync.
+	 *
+	 * <pre><code>
+	 * {
+	 *   Supported Versions: IntList
+	 * }
+	 * </code></pre>
+	 */
+	public record Handshake(IntList supportedVersions) implements FabricPacket {
+		public static final PacketType<Handshake> PACKET_TYPE = PacketType.create(ServerPackets.id("registry_sync/handshake_server"), Handshake::new);
+
+		public Handshake(FriendlyByteBuf buf) {
+			this(buf.readIntIdList());
+		}
+
+		@Override
+		public void write(FriendlyByteBuf buf) {
+			buf.writeIntIdList(this.supportedVersions);
+		}
+
+		@Override
+		public PacketType<Handshake> getType() {
+			return PACKET_TYPE;
+		}
+	}
+
+	/**
+	 * Ends registry sync. No data
+	 */
+	public record End() implements FabricPacket {
+		public static final PacketType<End> PACKET_TYPE = PacketType.create(ServerPackets.id("registry_sync/end"), End::new);
+
+		public End(FriendlyByteBuf buf) {
+			this();
+		}
+
+		@Override
+		public void write(FriendlyByteBuf buf) {
+		}
+
+		@Override
+		public PacketType<?> getType() {
+			return PACKET_TYPE;
+		}
+	}
+
+	/**
+	 * This packet sets failure text look/properties.
+	 * Requires protocol version 3 or newer.
+	 *
+	 * <pre><code>
+	 * {
+	 *   Text Header: Text (String)
+	 *   Text Footer: Text (String)
+	 *   Show Details: bool
+	 *
+	 * }
+	 * </code></pre>
+	 */
+	public record ErrorStyle(Component errorHeader, Component errorFooter, boolean showError) implements CustomPacketPayload {
+		public static final ResourceLocation ID = ServerPackets.id("registry_sync/error_style");
+
+		public ErrorStyle(FriendlyByteBuf buf) {
+			this(buf.readComponent(), buf.readComponent(), buf.readBoolean());
+		}
+
+		@Override
+		public void write(FriendlyByteBuf buf) {
+			buf.writeComponent(this.errorHeader);
+			buf.writeComponent(this.errorFooter);
+			buf.writeBoolean(this.showError);
+		}
+
+		@Override
+		public ResourceLocation id() {
+			return ID;
+		}
+	}
+
+	/**
+	 * This packet requests client to validate and return supported Mod Protocol versions.
+	 *
+	 * <pre><code>
+	 * {
+	 *   Prioritized Id: String
+	 *   Count of Entries: VarInt
+	 *   [
+	 *     Id: String
+	 *     Name: String
+	 *     Supported Versions: IntList
+	 *     Optional: boolean
+	 *   ]
+	 * }
+	 * </code></pre>
+	 */
+	public record ModProtocol(String prioritizedId, Collection<ModProtocolDef> protocols) implements CustomPacketPayload {
+		public static final ResourceLocation ID = ServerPackets.id("registry_sync/mod_protocol");
+
+		public ModProtocol(FriendlyByteBuf buf) {
+			this(buf.readUtf(), buf.readList(ModProtocolDef::read));
+		}
+
+		@Override
+		public void write(FriendlyByteBuf buf) {
+			buf.writeUtf(this.prioritizedId);
+			buf.writeCollection(this.protocols, ModProtocolDef::write);
+		}
+
+		@Override
+		public ResourceLocation id() {
+			return ID;
+		}
+	}
+
+	private static ResourceLocation id(String path) {
+		return FrozenMain.id(path);
+	}
+}