+++ src/main/java/org/quiltmc/qsl/frozenblock/core/registry/impl/sync/server/SyncTaskHolder.java
@@ -1,0 +1,6 @@
+package org.quiltmc.qsl.frozenblock.core.registry.impl.sync.server;
+
+public interface SyncTaskHolder {
+	QuiltSyncTask frozenLib$getQuiltSyncTask();
+	void frozenLib$finishQuiltSyncTask();
+}