+++ CHANGELOG.md
@@ -11,5 +11,8 @@ Put changelog here:
   - Configs will sync from server to client when joining a server, on datapack reload, and on use of `/frozenlib_config reload`
   - Configs will sync from client to server any time the client config is saved and the player has permission level 2 or greater
     - An example of this is saving a config via Cloth Config or other GUI
+- Added Mod Protocol API
+  - A partial implementation of The Quilt Project's Registry Sync API.
+  - Only 1.20.2+
 - Refactored Configured Features and Placement Modifiers
 - Removed embedded Mixin Extras in favor of Fabric Loader 0.15.0